tutorial finished

+++ app/Http/Controllers/ListingController.php
@@ -63,6 +63,10 @@ class ListingController extends Controller
     // Update Listing
     public function update(Request $request, Listing $listing)
     {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -85,7 +89,17 @@ class ListingController extends Controller
     // Delete Listing 
     public function destroy(Listing $listing)
     {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted succesfully');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
